# Count installed Scoop packages if available