--- a/Cargo.toml
+++ b/Cargo.toml
@@ -44,6 +44,7 @@ num_cpus = "1.13.0"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 local-ip-address = "0.4.4"
+which = "4.2.2"
 winreg = "0.10.1"
 windows = { version = "0.28.0", features = [
     "Win32_Foundation",