This change adds support for the [Scoop](https://scoop.sh/) installer on Windows.

It works by querying the `scoop export` command that lists all the installed packages and we can count the number of lines.