@grtcdr 2021-12-23
Lovely stuff... but I'm sure there's a directory whose entries we can count, right?