@FantasyTeddy 2021-12-23
Well... yes I think we could, but I don't know how to determine the correct path to Scoop itself. The default path is under `C:\Users\<user>\scoop` but it is possible to change this path during installation and if this is the case I'm not sure how to find the correct path...