@grtcdr 2021-12-23
Okay, how are the benchmarks looking?