@FantasyTeddy 2021-12-23
I tried to run the following benchmarks on my machine:
| Note | Average | Min | Max |
| - | - | - | - |
| Before this PR | 61.8 ms | 56.8 ms | 114.6 ms |
| PR without Scoop installed | 69.3 ms | 57.8 ms | 166.8 ms |
| PR with Scoop installed | 65.5 ms | 57.4 ms | 133.9 ms |