@FantasyTeddy 2021-12-23
Oh no, I just realized a mistake I made: even though I remembered to build `macchina` in release mode, I ran the following command to time it:
```
Measure-Command { .\target\debug\macchina.exe }
```
Can you spot my mistake?

---
I ran the tests again with the correct binary and now the results look very bad for my PR:
| Note | Average | Min | Max |
| - | - | - | - |
| Before this PR | 45.5 ms | 40.9 ms | 57.9 ms |
| PR without Scoop installed | 601.6 ms | 525.4 ms | 747.0 ms |
| PR with Scoop installed | 920.1 ms | 744.5 ms | 1178.7 ms |

I totally agree that this is not acceptable and I will have to find a way to determine the installation path of Scoop and do the directory enumeration.