@FantasyTeddy 2021-12-23
I tried to do it that way now: I'm looking for `scoop` with the [which](https://crates.io/crates/which) crate and find the correct path to the installed apps relative to this location.

The numbers look much better now, however there is still some impact (especially if the command for `scoop` cannot be found).
| Note | Average | Min | Max |
| - | - | - | - |
| Before this PR | 45.5 ms | 40.9 ms | 57.9 ms |
| PR without Scoop installed | 68.4 ms | 51.6 ms | 174.4 ms |
| PR with Scoop installed | 53.4 ms | 49.5 ms | 82.9 ms |