@grtcdr 2021-12-23
> Can you spot my mistake?

Measure-Command { .\target\**debug**\macchina.exe }

Yep.

The tests still show higher numbers than what is expected, can you please try benchmarking but with hyperfine instead? That's what we use for all our benchmarks.