@FantasyTeddy 2021-12-24
I measured it again with hyperfine and got these results:

| Note | Mean [ms] | Min [ms] | Max [ms] | Relative |
|---|---:|---:|---:|---:|
| Latest release (6.0.1) | 47.4 ± 3.1 | 42.7 | 62.6 | 1.00 |
| Latest `main` | 43.3 ± 1.1 | 41.0 | 46.4 | 1.00 |
| PR without Scoop installed | 54.0 ± 0.8 | 52.6 | 55.7 | 1.00 |
| PR with Scoop installed | 52.3 ± 1.7 | 49.5 | 60.2 | 1.00 |

I'm running an *Intel® Core™ i7-4770 CPU @ 3.40GHz* so it's not the fastest machine.