@rashil2000 2021-12-29
Sorry for bumping a merged thread, but a few days ago I wanted to add Scoop readout to macchina but couldn't get to it. Glad to see it's landed :)

> I'm looking for scoop with the which crate and find the correct path to the installed apps relative to this location.

How much cost does this dependency incur? I think we don't need it.

> Well... yes I think we could, but I don't know how to determine the correct path to Scoop itself. The default path is under `C:\Users\<user>\scoop` but it is possible to change this path during installation and if this is the case I'm not sure how to find the correct path...

The environment variable `SCOOP` is what you're looking for. See https://github.com/ScoopInstaller/Scoop#install-scoop-to-a-custom-directory-by-changing-scoop

If this variable is unset, Scoop defaults to `C:\Users\<user>\scoop`.

So if we can check this variable, we can slightly speed things up (I'm assuming an env var lookup is cheaper than searching for an executable in the PATH).