@FantasyTeddy 2021-12-29
> The environment variable `SCOOP` is what you're looking for. See https://github.com/ScoopInstaller/Scoop#install-scoop-to-a-custom-directory-by-changing-scoop
> 
> If this variable is unset, Scoop defaults to `C:\Users\<user>\scoop`.

I was looking at this too, but as far as I understand (and I could be totally wrong here) the `SCOOP` variable is only set temporarily to configure the installation, but it does not persist.
However I honestly did not try it and maybe it would work that way and then I am sure this would be a much faster alternative.
If I find some time I will happily try this and if it works I will rewrite it.