@rashil2000 2021-12-29
> I was looking at this too, but as far as I understand (and I could be totally wrong here) the `SCOOP` variable is only set temporarily to configure the installation, but it does not persist. 

If you see the PowerShell command there:
```powershell
[Environment]::SetEnvironmentVariable('SCOOP', $env:SCOOP, 'User')
```
this commands sets environment variables _permanently_ (as opposed to just the current shell session, like on unix).