@FantasyTeddy 2021-12-29
Ah that's the piece of information that was missing for me. Thank you very much for enlightening me :)
I will change the implementation accordingly and open another PR.